+++ lib/game/game_status.dart
@@ -1,0 +1,4 @@
+// Copyright © FullStackShack. All rights reserved.
+// Unauthorised use, reproduction, or distribution is strictly prohibited.
+
+enum GameStatus { playing, xWins, oWins, draw }